Math error fixes.
Forced the overlay height to equal 100 pixels greater than the height of
the document that way you wouldn't have a situation where for some
reason you scrolled out of it.

Fixed a bug where the arrow pointer position Y was being added to the
'top' calculation instead of being subtracted.  This resulted in the
image floating way to low for the 'N' arrow direction.

Defaulted the 'distanceBetween' variable to become 20 pixels if not
specified.  That's just a decision I made because I thought it would be
an alright distance regardless of what the user was looking for.
Obviously you can just override it if you want.