Fixing Firefox and IE bug
For some reason FadeIn on the exit tour image was causing IE and Firefox
to die before the tour started.  Now it just starts with a "show" since
it is in the first frame and will fade in with the main overlay.

Also moved the check to hide the exit image into if the tour has another
step to prevent extraneous calculation and make it easier to follow the
flow.